creates spectral ops test suite
