+++ test/test_torch.py
@@ -24,14 +24,14 @@ from random import randrange
 from torch import multiprocessing as mp
 from torch.testing._internal.common_methods_invocations import tri_tests_args, run_additional_tri_tests, \
     _compare_trilu_indices
-from torch.testing._internal.common_utils import TestCase, iter_indices, TEST_NUMPY, TEST_SCIPY, TEST_MKL, \
-    TEST_LIBROSA, TEST_WITH_ROCM, run_tests, skipIfNoLapack, suppress_warnings, \
+from torch.testing._internal.common_utils import TestCase, iter_indices, TEST_NUMPY, TEST_SCIPY, \
+    TEST_WITH_ROCM, run_tests, skipIfNoLapack, suppress_warnings, \
     IS_WINDOWS, NO_MULTIPROCESSING_SPAWN, do_test_dtypes, do_test_empty_full, \
     IS_SANDCASTLE, load_tests, slowTest, skipCUDANonDefaultStreamIf, skipCUDAMemoryLeakCheckIf, \
     BytesIOContext, skipIfRocm, torch_to_numpy_dtype_dict, skipIfNoSciPy, IS_MACOS, IS_PPC
 from multiprocessing.reduction import ForkingPickler
 from torch.testing._internal.common_device_type import instantiate_device_type_tests, \
-    skipCPUIfNoLapack, skipCPUIfNoMkl, skipCUDAIfNoMagma, skipCUDAIfRocm, skipCUDAIfNotRocm, onlyCUDA, onlyCPU, \
+    skipCPUIfNoLapack, skipCUDAIfNoMagma, skipCUDAIfRocm, skipCUDAIfNotRocm, onlyCUDA, onlyCPU, \
     dtypes, dtypesIfCUDA, dtypesIfCPU, deviceCountAtLeast, skipCUDAIf, precisionOverride, \
     PYTORCH_CUDA_MEMCHECK, largeCUDATensorTest, largeTensorTest, onlyOnCPUAndCUDA, expectedAlertNondeterministic
 from typing import Dict, List, Tuple, Union
@@ -50,9 +50,6 @@ if TEST_NUMPY:
 if TEST_SCIPY:
     import scipy
     from scipy import signal
-
-if TEST_LIBROSA:
-    import librosa
 
 SIZE = 100
 
@@ -1958,98 +1955,6 @@ class AbstractTestCases:
             torch.ormqr(m, tau, mat2, left=False, transpose=True, out=out_holder)
             self.assertEqual(res1, res2)
             self.assertEqual(res2, out_holder)
-
-        @staticmethod
-        def _test_fft_ifft_rfft_irfft(self, device='cpu', dtype=torch.double):
-            def _test_complex(sizes, signal_ndim, prepro_fn=lambda x: x):
-                x = prepro_fn(torch.randn(*sizes, dtype=dtype, device=device))
-                for normalized in (True, False):
-                    res = x.fft(signal_ndim, normalized=normalized)
-                    rec = res.ifft(signal_ndim, normalized=normalized)
-                    self.assertEqual(x, rec, atol=1e-8, rtol=0, msg='fft and ifft')
-                    res = x.ifft(signal_ndim, normalized=normalized)
-                    rec = res.fft(signal_ndim, normalized=normalized)
-                    self.assertEqual(x, rec, atol=1e-8, rtol=0, msg='ifft and fft')
-
-            def _test_real(sizes, signal_ndim, prepro_fn=lambda x: x):
-                x = prepro_fn(torch.randn(*sizes, dtype=dtype, device=device))
-                signal_numel = 1
-                signal_sizes = x.size()[-signal_ndim:]
-                for normalized, onesided in product((True, False), repeat=2):
-                    res = x.rfft(signal_ndim, normalized=normalized, onesided=onesided)
-                    if not onesided:  # check Hermitian symmetry
-                        def test_one_sample(res, test_num=10):
-                            idxs_per_dim = [torch.LongTensor(test_num).random_(s).tolist() for s in signal_sizes]
-                            for idx in zip(*idxs_per_dim):
-                                reflected_idx = tuple((s - i) % s for i, s in zip(idx, res.size()))
-                                idx_val = res.__getitem__(idx)
-                                reflected_val = res.__getitem__(reflected_idx)
-                                self.assertEqual(idx_val[0], reflected_val[0], msg='rfft hermitian symmetry on real part')
-                                self.assertEqual(idx_val[1], -reflected_val[1], msg='rfft hermitian symmetry on imaginary part')
-                        if len(sizes) == signal_ndim:
-                            test_one_sample(res)
-                        else:
-                            output_non_batch_shape = res.size()[-(signal_ndim + 1):]
-                            flatten_batch_res = res.view(-1, *output_non_batch_shape)
-                            nb = flatten_batch_res.size(0)
-                            test_idxs = torch.LongTensor(min(nb, 4)).random_(nb)
-                            for test_idx in test_idxs.tolist():
-                                test_one_sample(flatten_batch_res[test_idx])
-                        # compare with C2C
-                        xc = torch.stack([x, torch.zeros_like(x)], -1)
-                        xc_res = xc.fft(signal_ndim, normalized=normalized)
-                        self.assertEqual(res, xc_res)
-                    test_input_signal_sizes = [signal_sizes]
-                    rec = res.irfft(signal_ndim, normalized=normalized,
-                                    onesided=onesided, signal_sizes=signal_sizes)
-                    self.assertEqual(x, rec, atol=1e-8, rtol=0, msg='rfft and irfft')
-                    if not onesided:  # check that we can use C2C ifft
-                        rec = res.ifft(signal_ndim, normalized=normalized)
-                        self.assertEqual(x, rec.select(-1, 0), atol=1e-8, rtol=0, msg='twosided rfft and ifft real')
-                        self.assertEqual(rec.select(-1, 1).abs().mean(), 0, atol=1e-8,
-                                         rtol=0, msg='twosided rfft and ifft imaginary')
-
-            # contiguous case
-            _test_real((100,), 1)
-            _test_real((10, 1, 10, 100), 1)
-            _test_real((100, 100), 2)
-            _test_real((2, 2, 5, 80, 60), 2)
-            _test_real((50, 40, 70), 3)
-            _test_real((30, 1, 50, 25, 20), 3)
-
-            _test_complex((100, 2), 1)
-            _test_complex((100, 100, 2), 1)
-            _test_complex((100, 100, 2), 2)
-            _test_complex((1, 20, 80, 60, 2), 2)
-            _test_complex((50, 40, 70, 2), 3)
-            _test_complex((6, 5, 50, 25, 20, 2), 3)
-
-            # non-contiguous case
-            _test_real((165,), 1, lambda x: x.narrow(0, 25, 100))  # input is not aligned to complex type
-            _test_real((100, 100, 3), 1, lambda x: x[:, :, 0])
-            _test_real((100, 100), 2, lambda x: x.t())
-            _test_real((20, 100, 10, 10), 2, lambda x: x.view(20, 100, 100)[:, :60])
-            _test_real((65, 80, 115), 3, lambda x: x[10:60, 13:53, 10:80])
-            _test_real((30, 20, 50, 25), 3, lambda x: x.transpose(1, 2).transpose(2, 3))
-
-            _test_complex((2, 100), 1, lambda x: x.t())
-            _test_complex((100, 2), 1, lambda x: x.expand(100, 100, 2))
-            _test_complex((300, 200, 3), 2, lambda x: x[:100, :100, 1:])  # input is not aligned to complex type
-            _test_complex((20, 90, 110, 2), 2, lambda x: x[:, 5:85].narrow(2, 5, 100))
-            _test_complex((40, 60, 3, 80, 2), 3, lambda x: x.transpose(2, 0).select(0, 2)[5:55, :, 10:])
-            _test_complex((30, 55, 50, 22, 2), 3, lambda x: x[:, 3:53, 15:40, 1:21])
-
-            # non-contiguous with strides not representable as aligned with complex type
-            _test_complex((50,), 1, lambda x: x.as_strided([5, 5, 2], [3, 2, 1]))
-            _test_complex((50,), 1, lambda x: x.as_strided([5, 5, 2], [4, 2, 2]))
-            _test_complex((50,), 1, lambda x: x.as_strided([5, 5, 2], [4, 3, 1]))
-            _test_complex((50,), 2, lambda x: x.as_strided([5, 5, 2], [3, 3, 1]))
-            _test_complex((50,), 2, lambda x: x.as_strided([5, 5, 2], [4, 2, 2]))
-            _test_complex((50,), 2, lambda x: x.as_strided([5, 5, 2], [4, 3, 1]))
-
-        @unittest.skipIf(not TEST_MKL, "PyTorch is built without MKL support")
-        def test_fft_ifft_rfft_irfft(self):
-            self._test_fft_ifft_rfft_irfft(self)
 
         @unittest.skip("Not implemented yet")
         def test_conv2(self):
@@ -13839,327 +13744,6 @@ class TestTorchDeviceType(TestCase):
                         mean2 = x.mean(dim=dim, keepdim=keepdim)
                         self.assertEqual(var1, var2)
                         self.assertEqual(mean1, mean2)
-
-    # passes on ROCm w/ python 2.7, fails w/ python 3.6
-    @skipCUDAIfRocm
-    # stft -> rfft -> _fft -> _fft_with_size -> _fft_mkl
-    @skipCPUIfNoMkl
-    @dtypes(torch.double)
-    def test_stft(self, device, dtype):
-        if not TEST_LIBROSA:
-            raise unittest.SkipTest('librosa not found')
-
-        def librosa_stft(x, n_fft, hop_length, win_length, window, center):
-            if window is None:
-                window = np.ones(n_fft if win_length is None else win_length)
-            else:
-                window = window.cpu().numpy()
-            input_1d = x.dim() == 1
-            if input_1d:
-                x = x.view(1, -1)
-            result = []
-            for xi in x:
-                ri = librosa.stft(xi.cpu().numpy(), n_fft, hop_length, win_length, window, center=center)
-                result.append(torch.from_numpy(np.stack([ri.real, ri.imag], -1)))
-            result = torch.stack(result, 0)
-            if input_1d:
-                result = result[0]
-            return result
-
-        def _test(sizes, n_fft, hop_length=None, win_length=None, win_sizes=None,
-                  center=True, expected_error=None):
-            x = torch.randn(*sizes, dtype=dtype, device=device)
-            if win_sizes is not None:
-                window = torch.randn(*win_sizes, dtype=dtype, device=device)
-            else:
-                window = None
-            if expected_error is None:
-                result = x.stft(n_fft, hop_length, win_length, window, center=center)
-                # NB: librosa defaults to np.complex64 output, no matter what
-                # the input dtype
-                ref_result = librosa_stft(x, n_fft, hop_length, win_length, window, center)
-                self.assertEqual(result, ref_result, atol=7e-6, rtol=0, msg='stft comparison against librosa', exact_dtype=False)
-            else:
-                self.assertRaises(expected_error,
-                                  lambda: x.stft(n_fft, hop_length, win_length, window, center=center))
-
-        for center in [True, False]:
-            _test((10,), 7, center=center)
-            _test((10, 4000), 1024, center=center)
-
-            _test((10,), 7, 2, center=center)
-            _test((10, 4000), 1024, 512, center=center)
-
-            _test((10,), 7, 2, win_sizes=(7,), center=center)
-            _test((10, 4000), 1024, 512, win_sizes=(1024,), center=center)
-
-            # spectral oversample
-            _test((10,), 7, 2, win_length=5, center=center)
-            _test((10, 4000), 1024, 512, win_length=100, center=center)
-
-        _test((10, 4, 2), 1, 1, expected_error=RuntimeError)
-        _test((10,), 11, 1, center=False, expected_error=RuntimeError)
-        _test((10,), -1, 1, expected_error=RuntimeError)
-        _test((10,), 3, win_length=5, expected_error=RuntimeError)
-        _test((10,), 5, 4, win_sizes=(11,), expected_error=RuntimeError)
-        _test((10,), 5, 4, win_sizes=(1, 1), expected_error=RuntimeError)
-
-    @skipIfRocm
-    @unittest.skipIf(not TEST_MKL, "PyTorch is built without MKL support")
-    def test_fft_input_modification(self, device):
-        # FFT functions should not modify their input (gh-34551)
-
-        signal = torch.ones((2, 2, 2), device=device)
-        signal_copy = signal.clone()
-        spectrum = torch.fft(signal, 2)
-        self.assertEqual(signal, signal_copy)
-
-        spectrum_copy = spectrum.clone()
-        _ = torch.ifft(spectrum, 2)
-        self.assertEqual(spectrum, spectrum_copy)
-
-        half_spectrum = torch.rfft(signal, 2)
-        self.assertEqual(signal, signal_copy)
-
-        half_spectrum_copy = half_spectrum.clone()
-        _ = torch.irfft(half_spectrum_copy, 2, signal_sizes=(2, 2))
-        self.assertEqual(half_spectrum, half_spectrum_copy)
-
-    @onlyOnCPUAndCUDA
-    @unittest.skipIf(not TEST_MKL, "PyTorch is built without MKL support")
-    @dtypes(torch.double)
-    def test_istft_round_trip_simple_cases(self, device, dtype):
-        """stft -> istft should recover the original signale"""
-        def _test(input, n_fft, length):
-            stft = torch.stft(input, n_fft=n_fft)
-            inverse = torch.istft(stft, n_fft=n_fft, length=length)
-            self.assertEqual(input, inverse, exact_dtype=True)
-
-        _test(torch.ones(4, dtype=dtype, device=device), 4, 4)
-        _test(torch.zeros(4, dtype=dtype, device=device), 4, 4)
-
-    @onlyOnCPUAndCUDA
-    @unittest.skipIf(not TEST_MKL, "PyTorch is built without MKL support")
-    @dtypes(torch.double)
-    def test_istft_round_trip_various_params(self, device, dtype):
-        """stft -> istft should recover the original signale"""
-        def _test_istft_is_inverse_of_stft(stft_kwargs):
-            # generates a random sound signal for each tril and then does the stft/istft
-            # operation to check whether we can reconstruct signal
-            data_sizes = [(2, 20), (3, 15), (4, 10)]
-            num_trials = 100
-            istft_kwargs = stft_kwargs.copy()
-            del istft_kwargs['pad_mode']
-            for sizes in data_sizes:
-                for i in range(num_trials):
-                    original = torch.randn(*sizes, dtype=dtype, device=device)
-                    stft = torch.stft(original, **stft_kwargs)
-                    inversed = torch.istft(stft, length=original.size(1), **istft_kwargs)
-
-                    # trim the original for case when constructed signal is shorter than original
-                    original = original[..., :inversed.size(-1)]
-                    self.assertEqual(
-                        inversed, original, msg='istft comparison against original',
-                        atol=7e-6, rtol=0, exact_dtype=True)
-
-        patterns = [
-            # hann_window, centered, normalized, onesided
-            {
-                'n_fft': 12,
-                'hop_length': 4,
-                'win_length': 12,
-                'window': torch.hann_window(12, dtype=dtype, device=device),
-                'center': True,
-                'pad_mode': 'reflect',
-                'normalized': True,
-                'onesided': True,
-            },
-            # hann_window, centered, not normalized, not onesided
-            {
-                'n_fft': 12,
-                'hop_length': 2,
-                'win_length': 8,
-                'window': torch.hann_window(8, dtype=dtype, device=device),
-                'center': True,
-                'pad_mode': 'reflect',
-                'normalized': False,
-                'onesided': False,
-            },
-            # hamming_window, centered, normalized, not onesided
-            {
-                'n_fft': 15,
-                'hop_length': 3,
-                'win_length': 11,
-                'window': torch.hamming_window(11, dtype=dtype, device=device),
-                'center': True,
-                'pad_mode': 'constant',
-                'normalized': True,
-                'onesided': False,
-            },
-            # hamming_window, not centered, not normalized, onesided
-            # window same size as n_fft
-            {
-                'n_fft': 5,
-                'hop_length': 2,
-                'win_length': 5,
-                'window': torch.hamming_window(5, dtype=dtype, device=device),
-                'center': False,
-                'pad_mode': 'constant',
-                'normalized': False,
-                'onesided': True,
-            },
-            # hamming_window, not centered, not normalized, not onesided
-            # window same size as n_fft
-            {
-                'n_fft': 3,
-                'hop_length': 2,
-                'win_length': 3,
-                'window': torch.hamming_window(3, dtype=dtype, device=device),
-                'center': False,
-                'pad_mode': 'reflect',
-                'normalized': False,
-                'onesided': False,
-            },
-        ]
-        for i, pattern in enumerate(patterns):
-            _test_istft_is_inverse_of_stft(pattern)
-
-    @onlyOnCPUAndCUDA
-    def test_istft_throws(self, device):
-        """istft should throw exception for invalid parameters"""
-        stft = torch.zeros((3, 5, 2), device=device)
-        # the window is size 1 but it hops 20 so there is a gap which throw an error
-        self.assertRaises(
-            RuntimeError, torch.istft, stft, n_fft=4,
-            hop_length=20, win_length=1, window=torch.ones(1))
-        # A window of zeros does not meet NOLA
-        invalid_window = torch.zeros(4, device=device)
-        self.assertRaises(
-            RuntimeError, torch.istft, stft, n_fft=4, win_length=4, window=invalid_window)
-        # Input cannot be empty
-        self.assertRaises(RuntimeError, torch.istft, torch.zeros((3, 0, 2)), 2)
-        self.assertRaises(RuntimeError, torch.istft, torch.zeros((0, 3, 2)), 2)
-
-    @onlyOnCPUAndCUDA
-    @skipIfRocm
-    @skipCPUIfNoMkl
-    @dtypes(torch.double)
-    def test_istft_of_sine(self, device, dtype):
-        def _test(amplitude, L, n):
-            # stft of amplitude*sin(2*pi/L*n*x) with the hop length and window size equaling L
-            x = torch.arange(2 * L + 1, device=device, dtype=dtype)
-            original = amplitude * torch.sin(2 * math.pi / L * x * n)
-            # stft = torch.stft(original, L, hop_length=L, win_length=L,
-            #                   window=torch.ones(L), center=False, normalized=False)
-            stft = torch.zeros((L // 2 + 1, 2, 2), device=device, dtype=dtype)
-            stft_largest_val = (amplitude * L) / 2.0
-            if n < stft.size(0):
-                stft[n, :, 1] = -stft_largest_val
-
-            if 0 <= L - n < stft.size(0):
-                # symmetric about L // 2
-                stft[L - n, :, 1] = stft_largest_val
-
-            inverse = torch.istft(
-                stft, L, hop_length=L, win_length=L,
-                window=torch.ones(L, device=device, dtype=dtype), center=False, normalized=False)
-            # There is a larger error due to the scaling of amplitude
-            original = original[..., :inverse.size(-1)]
-            self.assertEqual(inverse, original, atol=1e-3, rtol=0)
-
-        _test(amplitude=123, L=5, n=1)
-        _test(amplitude=150, L=5, n=2)
-        _test(amplitude=111, L=5, n=3)
-        _test(amplitude=160, L=7, n=4)
-        _test(amplitude=145, L=8, n=5)
-        _test(amplitude=80, L=9, n=6)
-        _test(amplitude=99, L=10, n=7)
-
-    @onlyOnCPUAndCUDA
-    @skipIfRocm
-    @skipCPUIfNoMkl
-    @dtypes(torch.double)
-    def test_istft_linearity(self, device, dtype):
-        num_trials = 100
-
-        def _test(data_size, kwargs):
-            for i in range(num_trials):
-                tensor1 = torch.randn(data_size, device=device, dtype=dtype)
-                tensor2 = torch.randn(data_size, device=device, dtype=dtype)
-                a, b = torch.rand(2, dtype=dtype, device=device)
-                istft1 = torch.istft(tensor1, **kwargs)
-                istft2 = torch.istft(tensor2, **kwargs)
-                istft = a * istft1 + b * istft2
-                estimate = torch.istft(a * tensor1 + b * tensor2, **kwargs)
-                self.assertEqual(istft, estimate, atol=1e-5, rtol=0)
-        patterns = [
-            # hann_window, centered, normalized, onesided
-            (
-                (2, 7, 7, 2),
-                {
-                    'n_fft': 12,
-                    'window': torch.hann_window(12, device=device, dtype=dtype),
-                    'center': True,
-                    'normalized': True,
-                    'onesided': True,
-                },
-            ),
-            # hann_window, centered, not normalized, not onesided
-            (
-                (2, 12, 7, 2),
-                {
-                    'n_fft': 12,
-                    'window': torch.hann_window(12, device=device, dtype=dtype),
-                    'center': True,
-                    'normalized': False,
-                    'onesided': False,
-                },
-            ),
-            # hamming_window, centered, normalized, not onesided
-            (
-                (2, 12, 7, 2),
-                {
-                    'n_fft': 12,
-                    'window': torch.hamming_window(12, device=device, dtype=dtype),
-                    'center': True,
-                    'normalized': True,
-                    'onesided': False,
-                },
-            ),
-            # hamming_window, not centered, not normalized, onesided
-            (
-                (2, 7, 3, 2),
-                {
-                    'n_fft': 12,
-                    'window': torch.hamming_window(12, device=device, dtype=dtype),
-                    'center': False,
-                    'normalized': False,
-                    'onesided': True,
-                },
-            )
-        ]
-        for data_size, kwargs in patterns:
-            _test(data_size, kwargs)
-
-    @onlyOnCPUAndCUDA
-    @skipCPUIfNoMkl
-    @skipIfRocm
-    def test_batch_istft(self, device):
-        original = torch.tensor([
-            [[4., 0.], [4., 0.], [4., 0.], [4., 0.], [4., 0.]],
-            [[0., 0.], [0., 0.], [0., 0.], [0., 0.], [0., 0.]],
-            [[0., 0.], [0., 0.], [0., 0.], [0., 0.], [0., 0.]]
-        ], device=device)
-
-        single = original.repeat(1, 1, 1, 1)
-        multi = original.repeat(4, 1, 1, 1)
-
-        i_original = torch.istft(original, n_fft=4, length=4)
-        i_single = torch.istft(single, n_fft=4, length=4)
-        i_multi = torch.istft(multi, n_fft=4, length=4)
-
-        self.assertEqual(i_original.repeat(1, 1), i_single, atol=1e-6, rtol=0, exact_dtype=True)
-        self.assertEqual(i_original.repeat(4, 1), i_multi, atol=1e-6, rtol=0, exact_dtype=True)
 
     @skipCUDAIfRocm
     def test_blas_empty(self, device):
